HappyDog/backend/model-trainer/requirements.txt
torch
torchvision
Pillow
flask
numpy
pandas
scikit-learn==1.1.3